Extract mongo model to Schema object and create Model in different js file
