+++ server/db/mongooseConnection.js
@@ -1,0 +1,6 @@
+const mongoose = require('mongoose');
+
+// MONGO SETUP
+mongoose.set('debug', true);
+
+module.exports = mongoose.connect('mongodb://localhost/chat');
