+++ server/model/user.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+const Schema = mongoose.Schema;
+
+const userSchema = new Schema({
+  login: String,
+  password: String,
+  token: String,
+  tokenExpiredTime: Date
+});
+
+module.exports = mongoose.model('User', userSchema);
